# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gadget X Repairs API — lead capture, inventory browse, admin leads view.
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SubmitResponse {
  ok: boolean;
  id?: number;
}

export type ErrorResponseDetails = { [key: string]: unknown };

export interface ErrorResponse {
  error: string;
  details?: ErrorResponseDetails;
}

export type RepairQuoteInputPreferredContact =
  (typeof RepairQuoteInputPreferredContact)[keyof typeof RepairQuoteInputPreferredContact];

export const RepairQuoteInputPreferredContact = {
  call: "call",
  text: "text",
  whatsapp: "whatsapp",
  email: "email",
} as const;

export type RepairQuoteInputUrgency =
  (typeof RepairQuoteInputUrgency)[keyof typeof RepairQuoteInputUrgency];

export const RepairQuoteInputUrgency = {
  asap: "asap",
  today: "today",
  this_week: "this_week",
  flexible: "flexible",
} as const;

export interface RepairQuoteInput {
  /**
   * @minLength 1
   * @maxLength 120
   */
  name: string;
  /**
   * @minLength 7
   * @maxLength 40
   */
  phone: string;
  /** @maxLength 200 */
  email?: string;
  /**
   * @minLength 1
   * @maxLength 80
   */
  deviceType: string;
  /**
   * @minLength 1
   * @maxLength 80
   */
  brand: string;
  /**
   * @minLength 1
   * @maxLength 120
   */
  model: string;
  /**
   * @minLength 1
   * @maxLength 4000
   */
  problem: string;
  preferredContact?: RepairQuoteInputPreferredContact;
  urgency?: RepairQuoteInputUrgency;
  /** @maxLength 2000 */
  notes?: string;
  /** @maxLength 500 */
  photoUrl?: string;
}

export type SellPhoneInputLockedStatus =
  (typeof SellPhoneInputLockedStatus)[keyof typeof SellPhoneInputLockedStatus];

export const SellPhoneInputLockedStatus = {
  locked: "locked",
  unlocked: "unlocked",
} as const;

export type SellPhoneInputCondition =
  (typeof SellPhoneInputCondition)[keyof typeof SellPhoneInputCondition];

export const SellPhoneInputCondition = {
  mint: "mint",
  good: "good",
  fair: "fair",
  broken: "broken",
} as const;

export interface SellPhoneInput {
  /**
   * @minLength 1
   * @maxLength 120
   */
  name: string;
  /**
   * @minLength 7
   * @maxLength 40
   */
  phone: string;
  /**
   * @minLength 1
   * @maxLength 80
   */
  brand: string;
  /**
   * @minLength 1
   * @maxLength 120
   */
  model: string;
  /** @maxLength 40 */
  storage?: string;
  /** @maxLength 80 */
  carrier?: string;
  lockedStatus?: SellPhoneInputLockedStatus;
  condition?: SellPhoneInputCondition;
  /**
   * @minimum 0
   * @maximum 100
   */
  batteryHealth?: number;
  /** @maxLength 2000 */
  damageNotes?: string;
  /** @maxLength 20 */
  expectedPrice?: string;
  /** @maxLength 500 */
  photoUrl?: string;
}

export type AppointmentInputServiceType =
  (typeof AppointmentInputServiceType)[keyof typeof AppointmentInputServiceType];

export const AppointmentInputServiceType = {
  "screen-repair": "screen-repair",
  "battery-replacement": "battery-replacement",
  "hdmi-repair": "hdmi-repair",
  "laptop-diagnostic": "laptop-diagnostic",
  "motherboard-repair": "motherboard-repair",
  "phone-unlocking": "phone-unlocking",
  "prepaid-activation": "prepaid-activation",
} as const;

export interface AppointmentInput {
  /**
   * @minLength 1
   * @maxLength 120
   */
  name: string;
  /**
   * @minLength 7
   * @maxLength 40
   */
  phone: string;
  serviceType: AppointmentInputServiceType;
  /**
   * @minLength 1
   * @maxLength 80
   */
  preferredDatetime: string;
  /** @maxLength 2000 */
  notes?: string;
}

export interface ContactInput {
  /**
   * @minLength 1
   * @maxLength 120
   */
  name: string;
  /**
   * @minLength 3
   * @maxLength 200
   */
  contact: string;
  /**
   * @minLength 1
   * @maxLength 4000
   */
  message: string;
}

export interface ReservationInput {
  /**
   * @minLength 1
   * @maxLength 120
   */
  name: string;
  /**
   * @minLength 7
   * @maxLength 40
   */
  phone: string;
  /**
   * @minLength 1
   * @maxLength 120
   */
  itemId: string;
  /**
   * @minLength 1
   * @maxLength 200
   */
  itemLabel: string;
  /** @maxLength 1000 */
  notes?: string;
}

export interface InventoryItem {
  id: string;
  category: string;
  brand: string;
  model: string;
  storage?: string;
  color?: string;
  condition?: string;
  carrier?: string;
  price: string;
  warranty?: string;
  availability?: string;
  imageUrl?: string;
  description?: string;
}

export interface LeadBase {
  id: number;
  status: string;
  createdAt: string;
  updatedAt: string;
}

export type RepairQuoteRecord = LeadBase & RepairQuoteInput;

export type SellPhoneRecord = LeadBase & SellPhoneInput;

export type AppointmentRecord = LeadBase & AppointmentInput;

export type ContactRecord = LeadBase & ContactInput;

export type ReservationRecord = LeadBase & ReservationInput;

export type LeadStatusUpdateStatus =
  (typeof LeadStatusUpdateStatus)[keyof typeof LeadStatusUpdateStatus];

export const LeadStatusUpdateStatus = {
  new: "new",
  in_progress: "in_progress",
  done: "done",
  archived: "archived",
} as const;

export interface LeadStatusUpdate {
  status: LeadStatusUpdateStatus;
}

export interface MessagingConfig {
  emailEnabled: boolean;
  smsEnabled: boolean;
  mailFrom: string;
  smsFrom: string;
}

export type LeadCommunicationLeadType =
  (typeof LeadCommunicationLeadType)[keyof typeof LeadCommunicationLeadType];

export const LeadCommunicationLeadType = {
  "repair-quote": "repair-quote",
  "sell-phone": "sell-phone",
  appointment: "appointment",
  contact: "contact",
  reservation: "reservation",
} as const;

export type LeadCommunicationChannel =
  (typeof LeadCommunicationChannel)[keyof typeof LeadCommunicationChannel];

export const LeadCommunicationChannel = {
  email: "email",
  sms: "sms",
} as const;

export type LeadCommunicationDirection =
  (typeof LeadCommunicationDirection)[keyof typeof LeadCommunicationDirection];

export const LeadCommunicationDirection = {
  outbound: "outbound",
  inbound: "inbound",
} as const;

export interface LeadCommunication {
  id: number;
  leadType: LeadCommunicationLeadType;
  leadId: string;
  channel: LeadCommunicationChannel;
  direction: LeadCommunicationDirection;
  subject?: string | null;
  body: string;
  recipient: string;
  providerMessageId?: string | null;
  status: string;
  error?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface LeadActivityResponse {
  items: LeadCommunication[];
}

export interface SendEmailInput {
  /**
   * @minLength 3
   * @maxLength 200
   */
  to: string;
  /**
   * @minLength 1
   * @maxLength 200
   */
  subject: string;
  /** @maxLength 20000 */
  html?: string;
  /** @maxLength 20000 */
  text?: string;
}

export interface SendSmsInput {
  /**
   * @minLength 7
   * @maxLength 40
   */
  to: string;
  /**
   * @minLength 1
   * @maxLength 1600
   */
  body: string;
}

export type SendResultStatus =
  (typeof SendResultStatus)[keyof typeof SendResultStatus];

export const SendResultStatus = {
  queued: "queued",
  sent: "sent",
  failed: "failed",
} as const;

export interface SendResult {
  ok: boolean;
  id?: number;
  providerMessageId?: string | null;
  status: SendResultStatus;
  error?: string | null;
}

export interface WebhookAck {
  ok: boolean;
}

export interface AdminLeads {
  repairQuotes: RepairQuoteRecord[];
  sellPhoneSubmissions: SellPhoneRecord[];
  appointments: AppointmentRecord[];
  contactMessages: ContactRecord[];
  itemReservations: ReservationRecord[];
}

export type ListInventoryParams = {
  category?: string;
};