# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Gadget X Repairs API — lead capture, inventory browse, admin leads view.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /webhooks/resend:
    post:
      operationId: resendWebhook
      tags: [webhooks]
      summary: Resend delivery-status webhook (signature-verified)
      responses:
        "200":
          description: Acknowledged
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WebhookAck" }
        "401":
          description: Invalid signature
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /webhooks/telnyx:
    post:
      operationId: telnyxWebhook
      tags: [webhooks]
      summary: Telnyx delivery-status webhook (signature-verified)
      responses:
        "200":
          description: Acknowledged
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WebhookAck" }
        "401":
          description: Invalid signature
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
        details: { type: object, additionalProperties: true }
      required: [error]
...
    WebhookAck:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
...