# lib/api-zod/src/generated/types/webhookAck.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gadget X Repairs API — lead capture, inventory browse, admin leads view.
 * OpenAPI spec version: 0.1.0
 */

export interface WebhookAck {
  ok: boolean;
}